feat: move utils
